+++ adapters/prisma/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @proompter/adapter-prisma
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @proompter/core@0.0.4
+
+## 0.0.4
+
+### Patch Changes
+
+- use published flowise fork
+- Updated dependencies
+  - @proompter/core@0.0.3
+
 ## 0.0.3
 
 ### Patch Changes