+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @proompter/core
 
+## 0.0.4
+
+### Patch Changes
+
+- fix vercel ai versioning
+
+## 0.0.3
+
+### Patch Changes
+
+- use published flowise fork
+- Updated dependencies
+  - @proompter/ai@2.2.18
+
 ## 0.0.2
 
 ### Patch Changes